Hotel_Booking_Web/requirements.txt
flask
mysql.connector
flask_jwt_extended
flask_cors
dotenv
bcrypt
